requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.7
django-allauth==0.54.0
django-cors-headers==3.14.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
filelock==3.10.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
idna==2.8
oauthlib==3.2.2
Pillow==9.4.0
platformdirs==3.1.1
pycparser==2.21
PyJWT==1.7.1
pypng==0.20220715.0
python3-openid==3.2.0
pytz==2022.7.1
qrcode==7.4.2
ratelim==0.1.6
requests==2.22.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.25.11
virtualenv==20.21.0
